move tracker code to its own module
