+++ module/actor/actor-sheet.js
@@ -6,12 +6,12 @@ import { isConfigurationAllowed } from '../game-utils.js'
 import GurpsWiring from '../gurps-wiring.js'
 import { HitLocation, hitlocationDictionary } from '../hitlocation/hitlocation.js'
 import * as CI from '../injury/domain/ConditionalInjury.js'
+import { ResourceTrackerEditor } from '../resource-tracker/resource-tracker-editor.js'
+import { ResourceTrackerManager } from '../resource-tracker/resource-tracker-manager.js'
 import { Advantage, Equipment, Melee, Modifier, Note, Ranged, Reaction, Skill, Spell } from './actor-components.js'
 import { ActorImporter } from './actor-importer.js'
 import { cleanTags } from './effect-modifier-popout.js'
 import MoveModeEditor from './move-mode-editor.js'
-import { ResourceTrackerEditor } from './resource-tracker-editor.js'
-import { ResourceTrackerManager } from './resource-tracker-manager.js'
 import SplitDREditor from './splitdr-editor.js'
 
 /**
