+++ module/resource-tracker/resource-tracker-editor.js
